Vec changed to Vector

+++ contracts/grid/src/constants.rs
@@ -20,9 +20,9 @@ pub const MAX_PROTOCOL_FEE: u128 = 100000;
 pub const PROTOCOL_FEE_DENOMINATOR: u128 = 1000000;
 
 /// self.order_map[bot_id][FORWARD_ORDERS_INDEX]
-pub const FORWARD_ORDERS_INDEX: usize = 0;
+pub const FORWARD_ORDERS_INDEX: u64 = 0;
 /// self.order_map[bot_id][REVERSE_ORDERS_INDEX]
-pub const REVERSE_ORDERS_INDEX: usize = 1;
+pub const REVERSE_ORDERS_INDEX: u64 = 1;
 /// Forward and Reverse
-pub const ORDER_POSITION_SIZE: u16 = 2;
+pub const ORDER_POSITION_SIZE: u64 = 2;
